+++ Extensions/SessionExtension.cs
@@ -10,5 +10,16 @@ public static class SessionExtensions
             : default;
 
     public static void Set<T>(this Session session, string key, T value)
-        => session.Data[key] = value!;
+    {
+        session.Data[key] = value!;
+        session.MarkDirty();
+    }
+
+    public static bool Remove(this Session session, string key)
+    {
+        var removed = session.Data.TryRemove(key, out _);
+        if (removed)
+            session.MarkDirty();
+        return removed;
+    }
 }